set up inline-snapshot and test symmetry
